server: make configureMaxThreads platform-specific...
... and add a FreeBSD implementation
