+++ server/server.go
@@ -11,8 +11,6 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
-	"runtime/debug"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -363,23 +361,6 @@ func (s *Server) Shutdown(ctx context.Context) error {
 		close(s.ContainerEventsChan)
 	}
 
-	return nil
-}
-
-// configureMaxThreads sets the Go runtime max threads threshold
-// which is 90% of the kernel setting from /proc/sys/kernel/threads-max
-func configureMaxThreads() error {
-	mt, err := os.ReadFile("/proc/sys/kernel/threads-max")
-	if err != nil {
-		return fmt.Errorf("read max threads file: %w", err)
-	}
-	mtint, err := strconv.Atoi(strings.TrimSpace(string(mt)))
-	if err != nil {
-		return err
-	}
-	maxThreads := (mtint / 100) * 90
-	debug.SetMaxThreads(maxThreads)
-	logrus.Debugf("Golang's threads limit set to %d", maxThreads)
 	return nil
 }
